requirements.txt
yt-dlp
ffmpeg-python
PyYAML
openai-whisper
